refactor : Remove plugin to scripts
root 폴도 하위 plugin과 명칭이 중복되어 변경
