+++ dags/elt_dag.py
@@ -6,14 +6,13 @@ from airflow.operators.python import PythonOperator
 from airflow.providers.apache.spark.operators.spark_submit import SparkSubmitOperator
 from airflow.utils.dates import days_ago
 from airflow.utils.task_group import TaskGroup
-from plugins import slack
-from plugins.spark import get_etl_step, get_emr_cluster_id, get_elt_step
-from plugins.s3 import check_and_copy_files, upload_to_s3
 from airflow.providers.amazon.aws.operators.emr import EmrAddStepsOperator
 from airflow.providers.amazon.aws.sensors.emr import EmrStepSensor
 from airflow.providers.postgres.operators.postgres import PostgresOperator
 
-
+from dags.scripts.comm import slack
+from dags.scripts.comm.s3 import upload_to_s3
+from dags.scripts.comm.spark import get_elt_step, get_emr_cluster_id
 
 default_args = {
     'start_date': days_ago(1),
@@ -22,79 +21,72 @@ default_args = {
     'on_failure_callback': slack.send_failure_alert,
     'on_success_callback': slack.send_success_alert
 }
-script_key=f"emr/scripts/emr_elt_redshift.py"
+script_key = f"emr/scripts/emr_elt_redshift.py"
 bucket_name = "de-2-1-s3"
 spark_steps = get_elt_step(bucket_name, script_key)
 job_flow_name = 'de-2-1-emr'
 
-
-
 with DAG(
-    'emr_to_redshift',
-    default_args=default_args,
-    schedule_interval='30 * * * *',  # Run the DAG every 30 minutes past the hour
-    catchup=False,  # Do not catch up on historical runs
-)as dag:
-
-
-
+        'emr_to_redshift',
+        default_args=default_args,
+        schedule_interval='30 * * * *',  # Run the DAG every 30 minutes past the hour
+        catchup=False,  # Do not catch up on historical runs
+) as dag:
     with TaskGroup("truncate_table_job") as truncate_table_job:
-
         redshift_truncate_table_task1 = PostgresOperator(
             task_id='setup_create_table1',
             sql='TRUNCATE TABLE analytics.brand_basic_info;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task2 = PostgresOperator(
             task_id='setup_create_table2',
             sql='TRUNCATE TABLE analytics.brand_information;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task3 = PostgresOperator(
             task_id='setup_create_table3',
             sql='TRUNCATE TABLE analytics.aggregated_brand_information; ',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task4 = PostgresOperator(
             task_id='setup_create_table4',
             sql='TRUNCATE TABLE analytics.brand_media_post_time;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task5 = PostgresOperator(
             task_id='setup_create_table5',
             sql='TRUNCATE TABLE analytics.followers_growth; ',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task6 = PostgresOperator(
             task_id='setup_create_table6',
             sql='TRUNCATE TABLE analytics.aggregated_hashtag_search;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task7 = PostgresOperator(
             task_id='setup_create_table7',
             sql='TRUNCATE TABLE analytics.popularity_factor; ',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task8 = PostgresOperator(
             task_id='setup_create_table8',
             sql='TRUNCATE TABLE analytics.popularity_calculation;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
         redshift_truncate_table_task9 = PostgresOperator(
             task_id='setup_create_table9',
             sql='TRUNCATE TABLE analytics.hashtag_count;',
             postgres_conn_id="REDSHIFT_DEFAULT",
-            dag = dag
+            dag=dag
         )
-
 
     # spark_job
     get_cluster_id = PythonOperator(
